+++ src/modules/board/dto/board-create-request.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum, IsNumber, IsString } from "class-validator";
-import { StuffCategory } from "../../enums/stuffCategory.enum";
+import { StuffCategory } from "../../comment/enums/stuffCategory.enum";
 
 export class CreateBoardRequestDto {
